#CLOUD-494 fixed crash if using socket before mediator address has been retrieved
#CLOUD-494 fixed improper assert
#VMS-3422 added mediaserver<->cloud integration test helper classes. Added cloud authentication test
(VMS-3437) - light client can't view video
Internal: Minor code fixes and improvements.
Some changes in the testconf_local.py to run tests from jenkins.
Also data for testconf_local.py prepared and stored as a file instead of creating it by echoes.
Some warnings fixed.
New coding style applied to some files.
Minor improvements of debug/logging/test features.
VMS-2986
minor fix

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ appserver2/src/transaction/transaction_transport.cpp
@@ -719,7 +719,7 @@ void QnTransactionTransport::receivedTransactionNonSafe( const QnByteArrayConstR
         case Qn::JsonFormat:
             if( !QnJsonTransactionSerializer::deserializeTran(
                     reinterpret_cast<const quint8*>(tranData.constData()),
-                    tranData.size(),
+                    (int) tranData.size(),
                     transportHeader,
                     serializedTran ) )
             {
@@ -734,7 +734,7 @@ void QnTransactionTransport::receivedTransactionNonSafe( const QnByteArrayConstR
         case Qn::UbjsonFormat:
             if( !QnUbjsonTransactionSerializer::deserializeTran(
                     reinterpret_cast<const quint8*>(tranData.constData()),
-                    tranData.size(),
+                    (int) tranData.size(),
                     transportHeader,
                     serializedTran ) )
             {
